Fix test autoloading for FakePlayoffGenerator and CupTie import
FakePlayoffGenerator was nested inside PromotionRelegationValidationTest.php,
so PSR-4 couldn't autoload it when SeasonTransitionGuardTest referenced
Tests\Unit\FakePlayoffGenerator. Extract it to its own file. Also add the
missing CupTie import used by test_playoff_completed_uses_winner_not_next_position.
